some stuff i dont understnad that nic helped with
